In this project, I was asked to add test cases to the ones the professor gave to me, and program the elevator class to pass these test cases.
The elevator class were to simulate the operation of an elevator.  Details associated with opening and closing of the elevator doors, entry and exit of people to and from the elevator, the number of people in, entering or leaving the elevator, and the timing of the movement of the elevator are all "abstracted" out of the problem, encompassing all of these actions into a single "move()" operation.